+++ yampt.editor/source/editor_window.hpp
@@ -30,6 +30,7 @@ private:
 	plugin_workspace_view_t * m_plugin_workspace_view = nullptr;
 	settings_store_t m_settings { "yEditor.ini" };
 
+	QAction * m_conflicts_action = nullptr;
 	QToolButton * m_editing_btn = nullptr;
 	QLineEdit * m_search_field = nullptr;
 	QToolButton * m_case_sensitive_btn = nullptr;